// frontend/src/app/api/model/terminationWorkflowStatus.ts
/**
 * Atlas API
 *
 * Contact: TechSupport-ATLAS@sbb.ch
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type TerminationWorkflowStatus = 'STARTED' | 'TARIFF_STOP_APPROVED' | 'TARIFF_STOP_NOT_APPROVED' | 'TERMINATION_APPROVED' | 'TERMINATION_NOT_APPROVED' | 'TERMINATION_NOT_APPROVED_CLOSED' | 'CANCELED';

export const TerminationWorkflowStatus = {

    Started: 'STARTED' as TerminationWorkflowStatus,

    TariffStopApproved: 'TARIFF_STOP_APPROVED' as TerminationWorkflowStatus,

    TariffStopNotApproved: 'TARIFF_STOP_NOT_APPROVED' as TerminationWorkflowStatus,

    TerminationApproved: 'TERMINATION_APPROVED' as TerminationWorkflowStatus,

    TerminationNotApproved: 'TERMINATION_NOT_APPROVED' as TerminationWorkflowStatus,

    TerminationNotApprovedClosed: 'TERMINATION_NOT_APPROVED_CLOSED' as TerminationWorkflowStatus,

    Canceled: 'CANCELED' as TerminationWorkflowStatus
};