// frontend/src/app/api/model/terminationDecision.ts
/**
 * Atlas API
 *
 * Contact: TechSupport-ATLAS@sbb.ch
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { JudgementType } from './judgementType';


export interface TerminationDecision { 
    judgement: JudgementType;
    /**
     * Motivation
     */
    motivation?: string;
    /**
     * Decision by person: Info+ or NOVA
     */
    terminationDecisionPerson: TerminationDecision.TerminationDecisionPersonEnum;
    /**
     * Termination Date 
     */
    terminationDate?: Date;
    /**
     * Firstname
     */
    readonly firstName?: string;
    /**
     * Second
     */
    readonly lastName?: string;
    /**
     * Organisation
     */
    readonly organisation?: string;
    /**
     * Mail of examinant
     */
    readonly examinantMail?: string;
}
export namespace TerminationDecision {
    export type TerminationDecisionPersonEnum = 'INFO_PLUS' | 'NOVA';
    export const TerminationDecisionPersonEnum = {
        InfoPlus: 'INFO_PLUS' as TerminationDecisionPersonEnum,
        Nova: 'NOVA' as TerminationDecisionPersonEnum
    };
}