// frontend/src/app/api/model/terminationStopPointWorkflowModel.ts
/**
 * Atlas API
 *
 * Contact: TechSupport-ATLAS@sbb.ch
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TerminationDecision } from './terminationDecision';
import { TerminationWorkflowStatus } from './terminationWorkflowStatus';


export interface TerminationStopPointWorkflowModel { 
    /**
     * Workflow id
     */
    readonly id?: number;
    /**
     * Service Point version id
     */
    versionId: number;
    /**
     * Applicant mail
     */
    applicantMail: string;
    /**
     * Official designation
     */
    designationOfficial: string;
    /**
     * Unique code for locations that is used in customer information. The structure is described in the “Swiss Location ID” specification, chapter 4.2. The document is available here. https://transportdatamanagement.ch/standards/
     */
    sloid: string;
    status?: TerminationWorkflowStatus;
    /**
     * Workflow comment
     */
    workflowComment?: string;
    /**
     * Workflow abort comment
     */
    abortComment?: string;
    /**
     * StopPoint valid to
     */
    versionValidTo?: Date;
    /**
     * Termination Date defined by Business Organisation managing the Stop Point
     */
    boTerminationDate?: Date;
    /**
     * Termination Date defined by Info Plus
     */
    infoPlusTerminationDate?: Date;
    /**
     * Termination Date defined by NOVA
     */
    novaTerminationDate?: Date;
    infoPlusDecision?: TerminationDecision;
    novaDecision?: TerminationDecision;
    /**
     * Object creation date
     */
    creationDate?: string;
    /**
     * User creator
     */
    readonly creator?: string;
    /**
     * Last edition date
     */
    editionDate?: string;
    /**
     * Editor
     */
    readonly editor?: string;
}
export namespace TerminationStopPointWorkflowModel {
}